# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sunstar News API Documentation (सनस्टार न्युज)",
        "description": "Complete RESTful & Real-Time API for Sunstar News Portal: Articles, Live NEPSE & Rashifal Scrapers, Media Uploads/Downloads, Advertisement Banners, Dashboard Stats & Staff Management.",
        "contact": {
            "name": "Sunstar Tech Team",
            "email": "sitaram@sunstarnews.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Laravel Backend API Server"
        },
        {
            "url": "http://localhost:3001",
            "description": "Next.js Frontend Proxy Server"
        }
    ],
    "paths": {
...
        "/api/banners": {
            "get": {
                "tags": [
                    "Banners"
                ],
                "summary": "List all advertisement banners",
                "operationId": "9f718c7db0d13e4b5be26b80595f4cf1",
                "parameters": [
                    {
                        "name": "position",
                        "in": "query",
                        "description": "Filter by position (e.g. header-top, sidebar)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of banners",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Banners"
                ],
                "summary": "Create new banner advertisement",
                "operationId": "546707dc3442c51e694ec312df3664e9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "नयाँ विज्ञापन"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "header-top"
                                    },
                                    "targetUrl": {
                                        "type": "string",
                                        "example": "https://example.com"
                                    },
                                    "bannerFiles": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "imageUrl": {
                                        "type": "string",
                                        "example": "https://images.unsplash.com/photo-1557804506-669a67965ba0"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Banner created"
                    }
                }
            }
        },
        "/api/banners/{id}": {
            "delete": {
                "tags": [
                    "Banners"
                ],
                "summary": "Delete banner advertisement",
                "operationId": "00a27f787ec0124f0b98cfdeecf6090a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Banner ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Banner deleted"
                    }
                }
            }
        },
...
    },
...
}